search: distinguish "repo" and "repos" searches
The "repo" search engine will navigate directly to a specific
AudaxHealthInc (Rally) repository on GitHub by including the term in the
URL path, not a search query parameter.

The "repos" search engine has the prior behavior of "repo" before this
commit: it will search Rally repositories with the given terms.